+++ pkg/server/commands/listusers/list_users.go
@@ -110,8 +110,11 @@ type listUsersResponseMetadata struct {
 	// Atomic is used to be consistent with the Check and ListObjects.
 	DispatchCounter *atomic.Uint32
 
-	// WasThrottled indicates whether the request was throttled
-	WasThrottled *atomic.Bool
+	// WasDispatchThrottled indicates whether the request was dispatch throttled.
+	WasDispatchThrottled *atomic.Bool
+
+	// WasDatastoreThrottled indicates whether the request was datastore throttled.
+	WasDatastoreThrottled *atomic.Bool
 }
 
 func (r *listUsersResponse) GetUsers() []*openfgav1.User {